move sample parsing to mp4_sample_parser.py
